+++ source/extensions/omni.isaac.ros2_bridge/premake5.lua
@@ -216,8 +216,8 @@ repo_build.prebuild_copy {
 
 if os.target() == "linux" then
     repo_build.prebuild_copy {
-        { "%{root}/_build/target-deps/nv_ros2/lib/lib**", ext.target_dir.."/foxy/bin" },
-        { "%{root}/_build/target-deps/nv_ros2_humble/lib/lib**", ext.target_dir.."/humble/bin" },
+        { "%{root}/_build/target-deps/nv_ros2/lib/lib**", ext.target_dir.."/foxy/lib" },
+        { "%{root}/_build/target-deps/nv_ros2_humble/lib/lib**", ext.target_dir.."/humble/lib" },
         { "%{root}/_build/target-deps/nv_ros2/lib/python3.10/site-packages", ext.target_dir.."/foxy/rclpy" },
         { "%{root}/_build/target-deps/nv_ros2_humble/lib/python3.10/site-packages", ext.target_dir.."/humble/rclpy" },
         { "%{root}/_build/target-deps/nv_ros2_humble/local/lib/python3.10/dist-packages", ext.target_dir.."/humble/rclpy" },